+++ TeamCode/src/main/java/bifunctors/teleop/CompetitionTeleOp.java
@@ -5,9 +5,8 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import bifunctors.hardwaremap.HardwareMap;
 import bifunctors.helper.GamepadEx;
-import bifunctors.helper.Mecanum;
+import bifunctors.hardwaremap.components.Mecanum;
 import static bifunctors.helper.GamepadEx.primary;
-import static bifunctors.helper.GamepadEx.secondary;
 import static bifunctors.helper.GamepadEx.GamepadButton.*;
 
 @TeleOp(name="Competition", group="Linear OpMode")
